lab-3-spring-db-PR

+++ src/main/java/org/example/lab1/DTO/ProductDTO.java
@@ -7,19 +7,22 @@ import lombok.Builder;
 import lombok.Value;
 import org.example.lab1.validation.CosmicWordCheck;
 
+import java.math.BigDecimal;
+import java.util.UUID;
+
 @Value
 @Builder(toBuilder = true)
 public class ProductDTO {
-    long id;
+    Long id;
     @NotNull(message = "Product name cannot be null")
     @Size(min = 3, max = 100, message = "Product name must be between 3 and 100 characters")
     @CosmicWordCheck(message = "Product name must contain a cosmic term like 'star', 'galaxy', or 'comet'")
     String name;
     @NotNull(message = "Price cannot be null")
     @DecimalMin(value = "0.01", message = "Price must be greater than 0")
-    double price;
+    BigDecimal price;
     @NotNull(message = "Category ID cannot be null")
-    long categoryId;
+    Long categoryId;
     String description;
 
 }
